Calculate Distance Matrix

Calculates a distance matrix based on an object of class lsp.

lsp_to_dist(x, dist_fun, unit = "log2", p = NULL)

Arguments

x
An object of class lsp - usually the output of the lsp_thumbprint() function

dist_fun
A distance/dissimilarity method used. All possible values can be found using the philentropy::getDistMethods() function

unit
A character string specifying the logarithm unit that should be used to compute distances that depend on log computations: "log", "log2", "log10". The default is "log"

p
Power of the Minkowski distance. Used only when the dist_fun = "minkowski"

Value

An object of class `"dist"``

Examples

landcover = read_stars(system.file("raster/landcover2015.tif", package = "motif"))

landcover_cove = lsp_thumbprint(landcover, type = "cove", threshold = 0.9, window_size = 2000)
landcover_cove

#> # A tibble: 8 x 3
#>      id na_prop signature            
#> * <int>   <dbl> <list>               
#> 1     1       0 <dbl[,361] [1 × 361]>
#> 2     2       0 <dbl[,361] [1 × 361]>
#> 3     3       0 <dbl[,361] [1 × 361]>
#> 4     4       0 <dbl[,361] [1 × 361]>
#> 5     5       0 <dbl[,361] [1 × 361]>
#> 6     6       0 <dbl[,361] [1 × 361]>
#> 7     7       0 <dbl[,361] [1 × 361]>
#> 8     8       0 <dbl[,361] [1 × 361]>

dist_cov = lsp_to_dist(landcover_cove, dist_fun = "jensen-shannon")

#> Metric: 'jensen-shannon' using unit: 'log2'; comparing: 8 vectors.

dist_cov

#>           1         2         3         4         5         6         7
#> 2 1182978.9                                                            
#> 3  383101.1 1855868.8                                                  
#> 4 1728138.5 4218595.7 1437654.1                                        
#> 5 1986015.4 5088604.5 1619375.9  251836.4                              
#> 6  637309.5 1557323.0  425992.0 1723902.7 2074830.8                    
#> 7  805778.9  800180.0  791922.5 3106021.4 3855547.1  741990.0          
#> 8  730789.4 1694855.4  661207.7 1041069.8 1799698.9  555591.5  905331.3