motif: Local Pattern Analysis

Describes spatial patterns of categorical raster data for any defined regular and irregular areas. Patterns are described quantitatively using built-in signatures based on co-occurrence matrices but also allows for any user-defined functions. It enables spatial analysis such as search, change detection, and clustering to be performed on spatial patterns (Nowosad (2021) doi:10.1007/s10980-020-01135-0 ).

See also

Author

Maintainer: Jakub Nowosad nowosad.jakub@gmail.com (ORCID)