Prepares window* arguments (internal function)

Prepares window* arguments (internal function)

prepare_window(x, window)

Arguments

x

Object of class stars or stars_proxy

window

Specifies areas for analysis. It can be either: NULL, a numeric value, or an sf object. If window=NULL calculations are performed for a whole area. If the window argument is numeric, it is a length of the side of a square-shaped block of cells. Expressed in the numbers of cells, it defines the extent of a local pattern. If an sf object is provided, each feature (row) defines the extent of a local pattern. The sf object should have one attribute (otherwise, the first attribute is used as an id).

Value

A list with window, window_size, and window_shift