Changes structure of the lsp object

Converts a list-column with signatures into many numeric columns

lsp_restructure(x)

Arguments

x

an lsp object

Value

Object of class lsp. It has several columns: (1) id - an id of each window. For irregular windows, it is the values provided in the window argument, (2) na_prop - share (0-1) of NA cells for each window, (3) one or more columns representing values of the signature

Examples

library(stars)

landcover = read_stars(system.file("raster/landcover2015s.tif", package = "motif"))

landcover_cove = lsp_signature(landcover, type = "cove", threshold = 0.9, window = 100)
landcover_cover = lsp_restructure(landcover_cove)
landcover_cover
#> # A tibble: 47 × 51
#>       id na_prop      X1      X2      X3    X4    X5      X6    X7      X8    X9
#>  * <int>   <dbl>   <dbl>   <dbl>   <dbl> <dbl> <dbl>   <dbl> <dbl>   <dbl> <dbl>
#>  1     1    0    3.03e-4 6.82e-4 0           0     0 0           0 6.82e-4 0.998
#>  2     2    0    0       3.79e-4 0           0     0 0           0 3.79e-4 0.999
#>  3     3    0    2.16e-2 1.04e-2 7.58e-5     0     0 2.53e-5     0 1.04e-2 0.957
#>  4     4    0    6.10e-2 1.60e-2 0           0     0 5.56e-4     0 1.60e-2 0.906
#>  5     5    0    1.23e-1 2.99e-2 0           0     0 1.52e-4     0 2.99e-2 0.816
#>  6     6    0    1.26e-1 2.87e-2 0           0     0 1.59e-3     0 2.87e-2 0.804
#>  7     7    0.32 3.75e-2 1.28e-2 0           0     0 5.96e-4     0 1.28e-2 0.934
#>  8     8    0    1.22e-2 3.18e-3 3.23e-3     0     0 2.68e-3     0 3.18e-3 0.729
#>  9     9    0    1.85e-2 7.10e-3 7.83e-4     0     0 1.06e-3     0 7.10e-3 0.886
#> 10    10    0    7.07e-4 5.56e-4 2.53e-4     0     0 0           0 5.56e-4 0.922
#> # … with 37 more rows, and 40 more variables: X10 <dbl>, X11 <dbl>, X12 <dbl>,
#> #   X13 <dbl>, X14 <dbl>, X15 <dbl>, X16 <dbl>, X17 <dbl>, X18 <dbl>,
#> #   X19 <dbl>, X20 <dbl>, X21 <dbl>, X22 <dbl>, X23 <dbl>, X24 <dbl>,
#> #   X25 <dbl>, X26 <dbl>, X27 <dbl>, X28 <dbl>, X29 <dbl>, X30 <dbl>,
#> #   X31 <dbl>, X32 <dbl>, X33 <dbl>, X34 <dbl>, X35 <dbl>, X36 <dbl>,
#> #   X37 <dbl>, X38 <dbl>, X39 <dbl>, X40 <dbl>, X41 <dbl>, X42 <dbl>,
#> #   X43 <dbl>, X44 <dbl>, X45 <dbl>, X46 <dbl>, X47 <dbl>, X48 <dbl>, …

lsp_add_sf(landcover_cover)
#> Simple feature collection with 47 features and 51 fields
#> Geometry type: POLYGON
#> Dimension:     XY
#> Bounding box:  xmin: -400176.1 ymin: -609756.5 xmax: -190176.1 ymax: -399756.5
#> Projected CRS: unnamed
#> # A tibble: 47 × 52
#>       id na_prop      X1      X2      X3    X4    X5      X6    X7      X8    X9
#>    <int>   <dbl>   <dbl>   <dbl>   <dbl> <dbl> <dbl>   <dbl> <dbl>   <dbl> <dbl>
#>  1     1    0    3.03e-4 6.82e-4 0           0     0 0           0 6.82e-4 0.998
#>  2     2    0    0       3.79e-4 0           0     0 0           0 3.79e-4 0.999
#>  3     3    0    2.16e-2 1.04e-2 7.58e-5     0     0 2.53e-5     0 1.04e-2 0.957
#>  4     4    0    6.10e-2 1.60e-2 0           0     0 5.56e-4     0 1.60e-2 0.906
#>  5     5    0    1.23e-1 2.99e-2 0           0     0 1.52e-4     0 2.99e-2 0.816
#>  6     6    0    1.26e-1 2.87e-2 0           0     0 1.59e-3     0 2.87e-2 0.804
#>  7     7    0.32 3.75e-2 1.28e-2 0           0     0 5.96e-4     0 1.28e-2 0.934
#>  8     8    0    1.22e-2 3.18e-3 3.23e-3     0     0 2.68e-3     0 3.18e-3 0.729
#>  9     9    0    1.85e-2 7.10e-3 7.83e-4     0     0 1.06e-3     0 7.10e-3 0.886
#> 10    10    0    7.07e-4 5.56e-4 2.53e-4     0     0 0           0 5.56e-4 0.922
#> # … with 37 more rows, and 41 more variables: X10 <dbl>, X11 <dbl>, X12 <dbl>,
#> #   X13 <dbl>, X14 <dbl>, X15 <dbl>, X16 <dbl>, X17 <dbl>, X18 <dbl>,
#> #   X19 <dbl>, X20 <dbl>, X21 <dbl>, X22 <dbl>, X23 <dbl>, X24 <dbl>,
#> #   X25 <dbl>, X26 <dbl>, X27 <dbl>, X28 <dbl>, X29 <dbl>, X30 <dbl>,
#> #   X31 <dbl>, X32 <dbl>, X33 <dbl>, X34 <dbl>, X35 <dbl>, X36 <dbl>,
#> #   X37 <dbl>, X38 <dbl>, X39 <dbl>, X40 <dbl>, X41 <dbl>, X42 <dbl>,
#> #   X43 <dbl>, X44 <dbl>, X45 <dbl>, X46 <dbl>, X47 <dbl>, X48 <dbl>, …