Calculate Distance Matrix

Calculates a distance matrix based on an object of class lsp.

lsp_to_dist(x, dist_fun, unit = "log2", p = NULL)

Arguments

x
An object of class lsp - usually the output of the lsp_thumbprint() function

dist_fun
A distance/dissimilarity method used. All possible values can be found using the philentropy::getDistMethods() function

unit
A character string specifying the logarithm unit that should be used to compute distances that depend on log computations: "log", "log2", "log10". The default is "log"

p
Power of the Minkowski distance. Used only when the dist_fun = "minkowski"

Value

An object of class `"dist"``

Examples

library(stars)
landcover = read_stars(system.file("raster/landcover2015.tif", package = "motif"))

landcover_cove = lsp_thumbprint(landcover, type = "cove", threshold = 0.9, window_size = 2000)
landcover_cove

#> # A tibble: 6 x 3
#>      id na_prop signature          
#> * <int>   <dbl> <list>             
#> 1     1   0.659 <dbl[,49] [1 × 49]>
#> 2     2   0.374 <dbl[,49] [1 × 49]>
#> 3     3   0.705 <dbl[,49] [1 × 49]>
#> 4     6   0.637 <dbl[,49] [1 × 49]>
#> 5     7   0.458 <dbl[,49] [1 × 49]>
#> 6     8   0.682 <dbl[,49] [1 × 49]>

dist_cov = lsp_to_dist(landcover_cove, dist_fun = "jensen-shannon")

#> Metric: 'jensen-shannon' using unit: 'log2'; comparing: 6 vectors.

dist_cov

#>           1         2         3         6         7
#> 2  533977.8                                        
#> 3  246027.4 1011526.5                              
#> 6  473115.2 1154588.4  186837.0                    
#> 7  417059.7  468074.1  309643.4  393294.5          
#> 8  329342.9 1464061.0  207676.2  314121.5  864343.4